Task Deletion Authentication

+++ app/Http/routes.php
@@ -68,7 +68,12 @@ Route::group(['middleware' => ['web']], function () {
 	});
 
 	Route::delete('/task/{id}',function ($id) {
-	Task::findOrFail($id)->delete();
+
+        if(!Auth::check()){
+            return redirect("/task/$id")->withErrors(['You need to be authenticated to delete a task.']);
+
+        }
+        Task::findOrFail($id)->delete();
 
 	return redirect('/');
 	});
